dossier_Merwane/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mysql-connector-python==8.2.0
python-docx==1.1.0
python-multipart==0.0.6
pdfplumber
pytesseract
pdf2image
Pillow
PyPDF2==3.0.1
python-pptx
openpyxl